TERMINAR POSTMAN Y SUBIR

+++ database/migrations/2create_artists.php
@@ -16,7 +16,7 @@ return new class extends Migration {
         Schema::create('artists', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('followers')->default('0');
+            $table->unsignedBigInteger('followers')->default('0');
             $table->timestamps();
         });
     }
